Keep code DRY by extracting FloatingLogo component
