+++ src/pages/WalletManagement/ImportWalletRootPage.tsx
@@ -15,7 +15,6 @@
 // along with the library. If not, see <http://www.gnu.org/licenses/>.
 
 import { useState } from 'react'
-import styled from 'styled-components'
 
 import {
   WalletManagementContext,
@@ -27,9 +26,7 @@ import MultiStepsController from '../MultiStepsController'
 import ImportWordsPage from './ImportWordsPage'
 import WalletWelcomePage from './WalletWelcomePage'
 import AppHeader from '../../components/AppHeader'
-import { deviceBreakPoints } from '../../style/globalStyles'
-
-import { ReactComponent as AlephiumLogoSVG } from '../../images/alephium_logo_monochrome.svg'
+import FloatingLogo from '../../components/FloatingLogo'
 
 const ImportWalletRootPage = () => {
   const [context, setContext] = useState<WalletManagementContextType>(initialWalletManagementContext)
@@ -48,21 +45,5 @@ const ImportWalletRootPage = () => {
     </WalletManagementContext.Provider>
   )
 }
-
-const FloatingLogo = styled(AlephiumLogoSVG)`
-  position: absolute;
-  top: 50px;
-  left: 25px;
-  width: 60px;
-  height: 60px;
-
-  path {
-    fill: rgba(0, 0, 0, 0.05) !important;
-  }
-
-  @media ${deviceBreakPoints.mobile} {
-    display: none;
-  }
-`
 
 export default ImportWalletRootPage
